+++ CHANGELOG.md
@@ -6,10 +6,11 @@ features that will be removed in future versions **Removed** for deprecated feat
 
 ## Unreleased ##
 
-### v2.8.13 - QoS configuration
-  - **Update** Configuration of ROS quality of service by launchfile, #101
+### v2.8.14 - Laserscan messages for Multiscan
+  - **Update** Laserscan messages for Multiscan lidar, #96
 
-### v2.8.12 - Dynamical pointcloud transform
+### v2.8.13 - Dynamical pointcloud transform and QoS configuration
+  - **Update** Configuration of ROS quality of service by launchfile, #101
   - **Update** Dynamical configuration of an additional pointcloud transform by rosparam, #104
 
 ## Released ##